+++ .idea/misc.xml
@@ -7,7 +7,7 @@
         <entry key="app/src/main/res/layout/activity_main2.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/activity_menu.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/app_bar_main.xml" value="0.21979166666666666" />
-        <entry key="app/src/main/res/layout/fragment_albums.xml" value="0.3098958333333333" />
+        <entry key="app/src/main/res/layout/fragment_albums.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.3098958333333333" />
         <entry key="app/src/main/res/layout/nav_header_main.xml" value="0.1" />
         <entry key="app/src/main/res/menu/activity_main_drawer.xml" value="0.22083333333333333" />
