+++ app/src/main/java/com/example/vinilos/repositories/AlbumRepository.kt
@@ -3,11 +3,11 @@ package com.example.vinilos.repositories
 import android.app.Application
 import com.android.volley.VolleyError
 import com.example.vinilos.models.Album
-import com.example.vinilos.network.NetworkServiceAdapter
+import com.example.vinilos.network.AlbumServiceAdapter
 
 class AlbumRepository (val application: Application){
     fun refreshData(callback: (List<Album>)->Unit, onError: (VolleyError)->Unit) {
-        NetworkServiceAdapter.getInstance(application).getAlbums({
+        AlbumServiceAdapter.getInstance(application).getAlbums({
             callback(it)
         }, onError)
     }
