+++ app/src/main/res/values/colors.xml
@@ -12,4 +12,5 @@
     <color name="vinilos_green">#1ED760</color>
     <color name="vinilos_dark_gray">#212121</color>
     <color name="vinilos_light_gray">#484848</color>
+    <color name="mtrl_textinput_default_box_stroke_color">#FFFFFFFF</color>
 </resources>
